move core dir
