+++ .github/workflows/extension_ci.yml
@@ -69,7 +69,7 @@ jobs:
           sudo apt-get update && sudo apt-get install -y postgresql postgresql-contrib libopenblas-dev libreadline-dev
       - name: Test Core
         run: |
-          cd ../core && cargo test
+          cd ../server && cargo test
 
       - name: Test Core - Integration
         # skip when on external forks
@@ -81,7 +81,7 @@ jobs:
           PORTKEY_VIRTUAL_KEY_OPENAI: ${{ secrets.PORTKEY_VIRTUAL_KEY_OPENAI }}
           VOYAGE_API_KEY: ${{ secrets.VOYAGE_API_KEY }}
         run: |
-          cd ../core && cargo test -- --ignored
+          cd ../server && cargo test -- --ignored
 
   test:
     name: Run tests
