--- core/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "vectorize-core"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-crate-type = ["lib"]
-
-[dependencies]
-anyhow = "1.0.81"
-async-trait = "0.1.81"
-chrono = {version = "0.4.26", features = ["serde"] }
-env_logger = "0.11.3"
-lazy_static = "1.4.0"
-log = "0.4.21"
-ollama-rs = "=0.2.1"
-pgmq = "0.30"
-regex = "1.9.2"
-reqwest = {version = "0.11.18", features = ["json"] }
-serde = { version = "1.0.173", features = ["derive"] }
-serde_json = "1.0.103"
-sqlx = { version = "=0.8", features = [
-    "runtime-tokio-native-tls",
-    "postgres",
-    "chrono",
-    "json"
-] }
-thiserror = "1.0.44"
-tiktoken-rs = "0.5.7"
-tokio = {version = "1.29.1", features = ["rt-multi-thread"] }
-url = "2.5.0"
