+++ extension/src/api.rs
@@ -1,5 +1,7 @@
 use crate::chat::ops::{call_chat, call_chat_completions};
 use crate::chat::types::RenderedPrompt;
+use crate::core::query::{create_event_trigger, create_trigger_handler};
+use crate::core::types::{JobParams, Model};
 use crate::guc::get_guc_configs;
 use crate::init::{init_cron, VECTORIZE_QUEUE};
 use crate::search::{self, init_table};
@@ -8,8 +10,6 @@ use crate::transformers::transform;
 use crate::types;
 use crate::util::get_vectorize_meta_spi;
 use text_splitter::TextSplitter;
-use vectorize_core::query::{create_event_trigger, create_trigger_handler};
-use vectorize_core::types::{JobParams, Model};
 
 use anyhow::Result;
 use pgrx::prelude::*;
@@ -293,7 +293,7 @@ fn import_embeddings(
     let job_params: JobParams = serde_json::from_value(meta.params.clone())?;
 
     // Process rows based on table method
-    let count = if job_params.table_method == vectorize_core::types::TableMethod::join {
+    let count = if job_params.table_method == crate::core::types::TableMethod::join {
         let insert_q = format!(
             "INSERT INTO vectorize._embeddings_{} ({}, embeddings, updated_at)
              SELECT src.{}, src.{}, NOW()
