+++ server/Cargo.toml
@@ -13,21 +13,23 @@ name = "vectorize-worker"
 path = "src/bin/worker.rs"
 
 [dependencies]
-vectorize_core = { path = "../core", package = "vectorize-core"}
-
 actix-cors = "0.7.1"
 actix-http = "3.11.0"
 actix-rt = "2.10.0"
 actix-service = "2.0.3"
 actix-web = "4.11.0"
 anyhow = "1.0.98"
+async-trait = "0.1.88"
 chrono = {version = "0.4.41", features = ["serde"] }
 env = "1.0.1"
 env_logger = "0.11.8"
+lazy_static = "1.5.0"
 log = "0.4.27"
+ollama-rs = "=0.2.1"
 pgmq = "0.30.1"
 rand = "0.9.1"
-reqwest = "0.12.16"
+regex = "1.11.1"
+reqwest = { version = "0.12.16", features = ["json"] }
 serde = "1.0.219"
 serde_json = "1.0.140"
 sqlx = { version = "0.8", features = [ "runtime-tokio-native-tls", "postgres", "chrono", "json", "uuid"]}
