+++ server/src/core/errors.rs
