+++ server/src/core/transformers/providers/openai.rs
@@ -5,10 +5,10 @@ use super::{
     ChatMessageRequest, ChatResponse, EmbeddingProvider, GenericEmbeddingRequest,
     GenericEmbeddingResponse,
 };
-use crate::errors::VectorizeError;
-use crate::transformers::http_handler::handle_response;
-use crate::transformers::providers;
-use crate::transformers::types::Inputs;
+use crate::core::errors::VectorizeError;
+use crate::core::transformers::http_handler::handle_response;
+use crate::core::transformers::providers;
+use crate::core::transformers::types::Inputs;
 use async_trait::async_trait;
 use std::env;
