+++ server/src/errors.rs
@@ -1,4 +1,4 @@
-use vectorize_core::errors;
+use crate::core::errors;
 
 use actix_web::{HttpResponse, ResponseError, http::StatusCode, web::JsonConfig};
 use anyhow::Error as AnyhowError;
