+++ server/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
+pub mod core;
 pub mod db;
 pub mod errors;
 pub mod executor;
