--- vector-serve/Dockerfile.worker
@@ -1,12 +1,0 @@
-FROM rust:1.76.0-slim-buster as builder
-
-RUN apt-get update && \
-    apt-get install -y pkg-config libssl-dev && apt-get clean && \
-    rm -rf /var/lib/apt/lists/* /var/cache/apt/archives/*
-
-WORKDIR /build
-
-COPY core/ core/
-
-RUN cd core && \
-    cargo build --bin vectorize-worker --release
